+++ scripts/make_temp_models.py
@@ -69,6 +69,7 @@ tools = [
 
 bows = [
     'bow',
+    'crossbow',
     'slingshot'
 ]